qcacmn: change cfreq2 data type to int16_t
change cfreq2 data type from uint32_t to int16_t because its
deriving function return int16_t type value

CRs-Fixed: 3641069

+++ target_if/spectral/target_if_spectral.c
@@ -4468,7 +4468,7 @@ target_if_is_agile_span_overlap_with_operating_span
 	uint32_t op_end_freq;
 	uint32_t agile_start_freq;
 	uint32_t agile_end_freq;
-	uint32_t cfreq2;
+	int16_t cfreq2;
 
 	if (!spectral) {
 		spectral_err("spectral object is NULL");
